requirements.txt
snakemake>=8.0
pandas>=2.0
PyYAML>=6.0
requests>=2.31
tqdm>=4.66
Flask>=3.1.0